🔨 Restructure relation managers
